adding strict batch sizes
